feat: Move API to libs
